requirements.txt
# Gmail API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Google Sheets API
gspread==5.12.0
oauth2client==4.1.3

# Google Generative AI (Gemini)
google-generativeai==0.3.2

# Data processing and ML
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
scipy==1.11.4

# Text processing
nltk==3.8.1

# Database
sqlite3

# Email parsing
email-validator==2.1.0.post1

# Web scraping (if needed)
beautifulsoup4==4.12.2
requests==2.31.0

# Logging and configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development tools
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Jupyter notebooks (for analysis)
jupyter==1.0.0
ipykernel==6.27.1
matplotlib==3.8.2
seaborn==0.13.0